ai-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
numpy==1.26.2